# crate/server/documentation/openapi.yaml
...
openapi: 3.1.0
info:
  title: Cosmian KMS
  version: 5.0.0
  description: |
    High-performance, FIPS 140-3 compliant Key Management System implementing **KMIP 2.1 and 1.4** over HTTP/TLS.

    ## Authentication

    Most endpoints require one of the following authentication methods:

    - **Bearer JWT** — pass a JWT from your OIDC provider in `Authorization: Bearer <token>`
    - **API Token** — pass an API token in `Authorization: Bearer <token>` (server-issued opaque token)
    - **Mutual TLS (mTLS)** — present a client certificate during the TLS handshake

    Public endpoints (`/health`, `/version`, `/swagger`, `/openapi.yaml`) require no authentication.
servers:
  - url: /
    description: Current server
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT issued by your OIDC provider, or an opaque API token issued by the KMS server
    mtls:
      type: mutualTLS
      description: Client certificate presented during TLS handshake
    vaultToken:
      type: apiKey
      in: header
      name: X-Vault-Token
      description: |
        Vault-compatible bearer token issued by the Cosmian auth-verifier.
        Required on all `/v1/transit/*` and `/v1/pki/*` endpoints when
        `vault_api_enabled = true`.
    # ── REST Crypto tag management ───────────────────────────────────────────

    TagsRequest:
      description: Request body for adding or removing user tags on a key.
      type: object
      required: [tags]
      properties:
        tags:
          type: array
          items:
            type: string
          description: >
            Tags to add or remove. Each tag must be a non-empty string and must
            not start with `_` (that prefix is reserved for system tags).
          example: [jwks, prod]

    TagsResponse:
      description: Response body returned by all three tag management endpoints.
      type: object
      required: [kid, tags]
      properties:
        kid:
          type: string
          description: KMS unique identifier of the key.
...
paths:
...
  /server-info:
    get:
      tags: [Server]
      summary: Server configuration info
      description: Returns KMS version, FIPS mode, HSM instances, and the default username. Requires authentication.
      operationId: getServerInfo
      security:
        - bearerAuth: []
        - mtls: []
      responses:
        '200':
          description: Server info
          content:
            application/json:
              schema:
                type: object
                properties:
                  version:
                    type: string
                  fips_mode:
                    type: boolean
                  hsm_instances:
                    type: array
                    items:
                      type: object
                  default_username:
                    type: string
  /hsm/status:
    get:
      tags: [Server]
      summary: HSM status
      description: Returns the status of all configured HSM instances.
      operationId: getHsmStatus
      security:
        - bearerAuth: []
        - mtls: []
      responses:
        '200':
          description: HSM status list
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object

  # ── Certificates ─────────────────────────────────────────────────────────────
...